pyproject.toml
[project]
name = "spacealignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.7",
    "numpy>=2.3.3",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.3",
    "pandas-stubs>=2.3.2.250926",
    "pyquaternion>=0.9.9",
    "python-lsp-server>=1.13.1",
    "pytransform3d>=3.14.3",
    "rerun-sdk>=0.27.1",
    "scipy>=1.16.2",
    "scipy-stubs>=1.16.2.4",
]
